+++ packages/components/src/private-apis.ts
@@ -3,7 +3,7 @@ import { positionToPlacement as __experimentalPopoverLegacyPositionToPlacement }
 import { Menu } from './menu';
 import { ComponentsContext } from './context/context-system-provider';
 import { Tabs } from './tabs';
-import { kebabCase, normalizeTextString } from './utils/strings';
+import { normalizeTextString } from './utils/strings';
 import { lock } from './lock-unlock';
 import Badge from './badge';
 import { DateCalendar, DateRangeCalendar, TZDate } from './calendar';
@@ -30,7 +30,6 @@ lock( privateApis, {
 	ComponentsContext,
 	Tabs,
 	Menu,
-	kebabCase,
 	Badge,
 	normalizeTextString,
 	DateCalendar,